Fix: Do not delete summary node parent inside undo action
STDERR: Exception in thread "AWT-EventQueue-0" 
STDERR: java.lang.NullPointerException
STDERR: 	at
org.freeplane.features.map.mindmapmode.MMapController.deleteWithoutUndo(MMapController.java:326)
STDERR: 	at
org.freeplane.features.map.mindmapmode.MMapController$3.undo(MMapController.java:250)
STDERR: 	at
org.freeplane.core.undo.CompoundActor.undo(CompoundActor.java:71)
STDERR: 	at
org.freeplane.core.undo.UndoHandler.undo(UndoHandler.java:317)
STDERR: 	at
org.freeplane.core.undo.UndoHandler$UndoAction.actionPerformed(UndoHandler.java:47)

+++ freeplane/src/org/freeplane/features/map/SummaryNode.java
@@ -70,7 +70,7 @@ public class SummaryNode extends PersistentNodeHook implements IExtension{
 			
 			@Override
 			public void onNodeDeleted(NodeModel parent, NodeModel child, int index) {
-				if (! parent.isFolded() && ! parent.hasChildren() && isSummaryNode(parent)&& parent.getText().isEmpty()){
+				if (!modeController.isUndoAction() && ! parent.isFolded() && ! parent.hasChildren() && isSummaryNode(parent)&& parent.getText().isEmpty()){
 					MMapController mapController =  (MMapController) modeController.getMapController();
 					mapController.deleteNode(parent);
 				}
